src/streamlit/pyproject.toml
[tool.poetry]
name = "dsp-streamlit"
version = "0.1.0"
description = "DSP Streamlit package"
authors = ["Stijn Hering <stijn@hering.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.29.0"
plotly = "^5.18.0"
geopandas = "^0.14.1"
duckdb = "^0.9.2"
pygwalker = "^0.3.20"
polars = "^0.20.3"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
